requirements.txt
# for widgets.office
gspread==3.1.0
oauth2client==4.1.3
PyOpenSSL==19.0.0
python-dateutil==2.8.1

# for widgets.commute
googlemaps==3.1.3

# for widgets.room
requests==2.22.0

# for widgets.smog
bs4==0.0.1

# for widgets.allergy
py-air-control==0.5.0

# for widgets.calendar
cachetools==3.1.1
google-api-python-client==1.7.11
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1

# for epd* and display
# note: compiling Pillow from source needs "sudo apt-get install libjpeg-dev zlib1g-dev"
Pillow==4.0.0
RPi.GPIO==0.6.5
spidev==3.4
# note: Numpy needs sudo apt-get install python-dev libatlas-base-dev
numpy==1.17.3